+++ goko/Cargo.toml
@@ -34,8 +34,9 @@ protobuf = "2.16.2"
 rand = "0.7.3"
 yaml-rust = "0.4.4"
 pbr = "1.0.3"
+fxhash = "0.2.1"
 rayon = "1.3.1"
-indexmap = {version = "1.4.0", features = ["serde-1"]}
+hashbrown = { version = "0.8", features = ["rayon"] }
 crossbeam-channel = "0.4.2"
 pointcloud = { version = "0.3.0", path = "../pointcloud" }
 #evmap = { git = "https://github.com/comath/rust-evmap" }
