+++ goko/src/covertree/query_tools/knn_query_heap.rs
