+++ goko/src/covertree/query_tools/trace_query_heap.rs
