+++ goko/src/lib.rs
@@ -62,19 +62,17 @@ pub use errors::GokoResult;
 
 pub(crate) mod evmap;
 
-mod builders;
-mod data_caches;
-pub mod layer;
-pub mod node;
-pub mod query_tools;
-mod tree;
+mod covertree;
+pub use covertree::builders::CoverTreeBuilder;
+pub use covertree::*;
+
+pub mod query_interface;
+
 mod tree_file_format;
 pub mod utils;
 
 pub mod plugins;
 
-pub use builders::CoverTreeBuilder;
-pub use tree::*;
 
 /// The data structure explicitly seperates the covertree by layer, and the addressing schema for nodes
 /// is a pair for the layer index and the center point index of that node.
